accel/amdxdna: add the shared aie4 header (aie4.h)
Redo of the aie4 query/policy split so a platform (non-PCI) transport can share
the aie4 core. Make aie4.h the single home for the transport-neutral aie4
device/hwctx state and prototypes: aie4_pci.h now includes aie4.h and keeps only
the PCI-specific declarations (amdxdna_dev_priv, the SR-IOV helpers and the
aie4_pf/vf/classic_ops tables), and aie4_pci.c's query/policy and debugfs helpers
only lose the 'static' keyword so they match the shared prototypes (no code is
moved).

Point the transport-neutral aie4_ctx.c and aie4_error.c at aie4.h instead of
aie4_pci.h so they no longer drag in the PCI-only header (and <linux/pci.h>).

The transport-independent implementation (aie4.c, aie4_debugfs.c, aie4_dpt.c) is
added in the next commit. Nothing built changes here; the default PCI build is
unaffected.
