+++ drivers/accel/amdxdna/aie4_ctx.c
@@ -16,8 +16,8 @@
 #include <linux/types.h>
 
 #include "aie.h"
+#include "aie4.h"
 #include "aie4_host_queue.h"
-#include "aie4_pci.h"
 #include "aie4_msg_priv.h"
 #include "amdxdna_ctx.h"
 #include "amdxdna_gem.h"
