+++ drivers/accel/amdxdna/aie4_error.c
@@ -17,8 +17,8 @@
 #include <linux/workqueue.h>
 
 #include "aie.h"
+#include "aie4.h"
 #include "aie4_msg_priv.h"
-#include "aie4_pci.h"
 #include "amdxdna_ctx.h"
 #include "amdxdna_error.h"
 #include "amdxdna_pci_drv.h"
